move utility methods to utils package
